解决网关递归后出现空指针异常

+++ rear/advertise/src/main/java/com/spaceobj/advertise/bo/JdAdvertisBo.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 @Builder
 public class JdAdvertisBo implements Serializable {
 
-    private Long jdAdId;
+    private Integer jdAdId;
 
     private String jdAdHyperlink;
